// src/types/spotifyTypes.ts
export interface TrackParams {
    seed_genres: string;
    target_tempo: string;
    target_energy: string;
    target_acousticness: string;
    target_danceability: string;
    target_instrumentalness: string;
    target_liveness: string;
    target_speechiness: string;
    target_popularity: string;
    market: string;
    limit: string;
    offset: string;
  }